refactor: Move uploadFile function to a separate file for better organization
